+++ services/device-registry-base/src/main/java/org/eclipse/hono/service/management/credentials/CredentialsDto.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 Contributors to the Eclipse Foundation
+ * Copyright (c) 2020, 2021 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
-package org.eclipse.hono.deviceregistry.mongodb.model;
+package org.eclipse.hono.service.management.credentials;
 
 import java.net.HttpURLConnection;
 import java.time.Instant;
@@ -22,23 +22,29 @@ import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import org.eclipse.hono.client.ClientErrorException;
-import org.eclipse.hono.deviceregistry.mongodb.utils.MongoDbDeviceRegistryUtils;
 import org.eclipse.hono.service.management.BaseDto;
-import org.eclipse.hono.service.management.credentials.CommonCredential;
 import org.eclipse.hono.util.RegistryManagementConstants;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
- * A DTO (Data Transfer Object) class to store credentials information in mongodb.
+ * A Data Transfer Object for credentials information.
+ * <p>
+ * This is basically a wrapper around a list of {@link CommonCredential} objects, adding a resource version
+ * and time stamps for initial creation and last update.
  */
 public final class CredentialsDto extends BaseDto<List<CommonCredential>> {
+
+    /**
+     * The name of the JSON property containing the credentials.
+     */
+    public static final String FIELD_CREDENTIALS = "credentials";
 
     @JsonProperty(value = RegistryManagementConstants.FIELD_PAYLOAD_TENANT_ID, required = true)
     private String tenantId;
 
-    @JsonProperty(value = RegistryManagementConstants.FIELD_PAYLOAD_DEVICE_ID, required = true)
+    @JsonProperty(value = RegistryManagementConstants.FIELD_PAYLOAD_DEVICE_ID)
     private String deviceId;
 
     private boolean requiresMerging;
@@ -50,24 +56,33 @@ public final class CredentialsDto extends BaseDto<List<CommonCredential>> {
     }
 
     /**
-     * Constructs a new DTO for use with the <b>creation of a new</b> persistent entry.
+     * Creates a DTO for persisting credentials.
      * <p>
-     * This constructor also makes sure that the identifiers of the credentials'
+     * This method also makes sure that the identifiers of the credentials'
      * secrets are unique within their credentials object.
      * <p>
      * These properties are supposed to be asserted before persisting
      *
-     * @param tenantId The tenant identifier.
-     * @param deviceId The device identifier.
-     * @param credentials The list of credentials.
-     * @param version The version of the credentials to be sent as request header.
+     * @param tenantId The identifier of the tenant that the device belongs to.
+     * @param deviceId The identifier of the device that the credentials belong to.
+     * @param credentials The list of credentials to store.
+     * @param version The credentials' resource version or {@code null} if the resource version should be created
+     *                automatically.
      *
-     * @return A DTO instance for creating a new entry.
+     * @return The DTO.
      *
-     * @throws NullPointerException if any of the parameters except credentials is {@code null}
-     * @throws org.eclipse.hono.client.ClientErrorException if any of the checks fail.
+     * @throws NullPointerException if any of the parameters are {@code null}.
+     * @throws ClientErrorException if any of the credentials checks fail.
      */
-    public static CredentialsDto forCreation(final String tenantId, final String deviceId, final List<CommonCredential> credentials, final String version) {
+    public static CredentialsDto forCreation(
+            final String tenantId,
+            final String deviceId,
+            final List<CommonCredential> credentials,
+            final String version) {
+
+        Objects.requireNonNull(tenantId);
+        Objects.requireNonNull(deviceId);
+
         final CredentialsDto credentialsDto = BaseDto.forCreation(CredentialsDto::new, credentials, version);
         credentialsDto.setTenantId(tenantId);
         credentialsDto.setDeviceId(deviceId);
@@ -76,16 +91,65 @@ public final class CredentialsDto extends BaseDto<List<CommonCredential>> {
     }
 
     /**
-     * Constructs a new DTO for use with the <b>updating</b> a persistent entry.
+     * Creates a DTO for credentials read from the persistent store.
      *
-     * @param tenantId The id of the tenant.
-     * @param deviceId The id of the device.
-     * @param credentials The data of the DTO.
-     * @param version The version of the DTO
+     * @param tenantId The identifier of the tenant that the device belongs to.
+     * @param deviceId The identifier of the device that the credentials belong to.
+     * @param credentials The list of credentials from the store.
+     * @param created The point in time when the credentials were created initially in the store (may be {@code null}).
+     * @param updated The point in time when the credentials were updated most recently in the store (may be {@code null}).
+     * @param version The resource version of the credentials in the store.
      *
-     * @return A DTO instance for updating an entry.
+     * @return The DTO.
+     * @throws NullPointerException if tenant ID, device ID or credentials are {@code null}.
      */
-    public static CredentialsDto forUpdate(final String tenantId, final String deviceId, final List<CommonCredential> credentials, final String version) {
+    public static CredentialsDto forRead(
+            final String tenantId,
+            final String deviceId,
+            final List<CommonCredential> credentials,
+            final Instant created,
+            final Instant updated,
+            final String version) {
+
+        Objects.requireNonNull(tenantId);
+        Objects.requireNonNull(deviceId);
+
+        final CredentialsDto credentialsDto = BaseDto.forRead(
+                CredentialsDto::new,
+                credentials,
+                created,
+                updated,
+                version);
+        credentialsDto.setTenantId(tenantId);
+        credentialsDto.setDeviceId(deviceId);
+
+        return credentialsDto;
+    }
+
+    /**
+     * Creates a DTO for updating credentials.
+     * <p>
+     * This method also makes sure that the identifiers of the credentials'
+     * secrets are unique within their credentials object.
+     *
+     * @param tenantId The identifier of the tenant that the device belongs to.
+     * @param deviceId The identifier of the device that the credentials belong to.
+     * @param credentials The list of credentials to store.
+     * @param version The resource version of the object in the store to be updated.
+     *
+     * @return The DTO.
+     * @throws NullPointerException if tenant ID, device ID or credentials are {@code null}.
+     * @throws ClientErrorException if any of the credentials checks fail.
+     */
+    public static CredentialsDto forUpdate(
+            final String tenantId,
+            final String deviceId,
+            final List<CommonCredential> credentials,
+            final String version) {
+
+        Objects.requireNonNull(tenantId);
+        Objects.requireNonNull(deviceId);
+
         final CredentialsDto credentialsDto = BaseDto.forUpdate(CredentialsDto::new, credentials, version);
         credentialsDto.setTenantId(tenantId);
         credentialsDto.setDeviceId(deviceId);
@@ -165,13 +229,15 @@ public final class CredentialsDto extends BaseDto<List<CommonCredential>> {
 
     @Override
     protected void setData(final List<CommonCredential> data) {
-        //Validate the given credentials, secrets and generate secret ids if not available.
-        Optional.ofNullable(data)
-                .ifPresent(creds -> {
-                    assertSecretIdUniqueness(creds);
-                });
-
+        // validate the given credentials, secrets and generate secret ids if not available.
+        Optional.ofNullable(data).ifPresent(this::assertSecretIdUniqueness);
         super.setData(data);
+    }
+
+    @Override
+    @JsonProperty(FIELD_CREDENTIALS)
+    public List<CommonCredential> getData() {
+        return super.getData();
     }
 
     /**
@@ -198,25 +264,30 @@ public final class CredentialsDto extends BaseDto<List<CommonCredential>> {
     }
 
     /**
-     * Merges the secrets of the given credential DTO with that of the current one.
+     * Merges the secrets of the given credential DTO with those of this one.
      *
-     * @param otherCredentialsDto The credential DTO to be merged.
-     * @return  a reference to this for fluent use.
+     * @param otherCredentialsDto The DTO to be merged into this one.
+     * @return A reference to this for fluent use.
      * @throws NullPointerException if the given credentials DTO is {@code null}.
+     * @throws IllegalArgumentException if the given credentials DTO does not contain any credentials
      */
     @JsonIgnore
     public CredentialsDto merge(final CredentialsDto otherCredentialsDto) {
         Objects.requireNonNull(otherCredentialsDto);
 
-        Optional.ofNullable(otherCredentialsDto.getCredentials())
-                .ifPresent(credentialsToMerge -> this.getData()
-                        .forEach(credential -> findCredentialByIdAndType(credential.getAuthId(), credential.getType(),
-                                credentialsToMerge)
-                                        .ifPresent(credential::merge)));
+        final var credentialsToMerge = otherCredentialsDto.getCredentials();
+        if (credentialsToMerge == null || credentialsToMerge.isEmpty()) {
+            throw new IllegalArgumentException("no credentials on record for device");
+        } else {
+            this.getData().forEach(credential -> findCredentialByIdAndType(
+                    credential.getAuthId(),
+                    credential.getType(),
+                    credentialsToMerge).ifPresent(credential::merge));
 
-        setUpdatedOn(Instant.now());
+            setUpdatedOn(Instant.now());
+            return this;
+        }
 
-        return this;
     }
 
     @JsonIgnore
@@ -228,11 +299,5 @@ public final class CredentialsDto extends BaseDto<List<CommonCredential>> {
         return credentials.stream()
                 .filter(credential -> authId.equals(credential.getAuthId()) && authType.equals(credential.getType()))
                 .findFirst();
-    }
-
-    @Override
-    @JsonProperty(value = MongoDbDeviceRegistryUtils.FIELD_CREDENTIALS, required = true)
-    public List<CommonCredential> getData() {
-        return super.getData();
     }
 }